+++ src/index.ts
@@ -14,7 +14,12 @@ export {
 
 export * from "./core/mst-operations"
 export { escapeJsonPath, unescapeJsonPath, IJsonPatch } from "./core/json-patch"
-export { onAction } from "./core/action"
+export {
+    onAction,
+    IMiddleWareEvent,
+    IMiddlewareEventType,
+    ISerializedActionCall
+} from "./core/action"
 export { process } from "./core/process"
 export { isStateTreeNode, IStateTreeNode } from "./core/node"